% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCytoData.R
\name{createCytoData}
\alias{createCytoData}
\title{Read the cytotoxicity data for a given cyto_type (AB or LDH) from an input data chunk for a single plate}
\usage{
createCytoData(cyto_data, cyto_type, Plate.SN = NULL, date = NULL)
}
\arguments{
\item{cyto_data}{data chunk from a file containing the chemical, concentration, and blank-corrected values for a single assay and plate}

\item{cyto_type}{type of cytotoxicity data in the input cyto_data (either AB or LDH)}

\item{Plate.SN}{plate serial number of plate represented in cyto_data}

\item{date}{culture data of data represented in cyto_data}
}
\description{
Read the cytotoxicity data for a given cyto_type (AB or LDH) from an input data chunk for a single plate
}
\details{
The chemical, concentration, and blank-corrected data values are read. The following process is used to identify the indicies in the input data chunk that
contain the desired data:
\itemize{
\item Chemical data =
}
\enumerate{
\item Search for word “Chemical”
\item Search for word “Row” in same column as “Chemical”
\item Take all data that is 1 – 6 rows below the word “Row” and in the same column as “Chemical” to + 8 rows out
}
\itemize{
\item Concentration data =
Same as previous, except replace “Chemical” with “Concentration mM”
\item Blank-corrected values =
}
\enumerate{
\item Same as previous, except “Chemical” is replaced with one of the following: c("Corrected for Blank", "Corrected Optical Denisty 490 nm", "Corrected Optical Density 490 nm", "Corrected Fluorescence")
}

Note that any negative blank-corrected values are set to 0.
}
