% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCytoTable2.R
\name{createCytoTable2}
\alias{createCytoTable2}
\title{Create a table of the cytotoxicity data for the given cyto_type (AB or LDH) for all plates in a given input cytoFile}
\usage{
createCytoTable2(cytoFile, cyto_type)
}
\arguments{
\item{cytoFile}{full path to single .xlsx file containing the blank-corrected cytotoxicity data values (either a "Calculations" file, which is assumed to contain data for 3 plates,
or a "Summary" file, which is assumed to contain data for 1 plate)}

\item{cyto_type}{type of cytotoxicity data to read (either AB or LDH)}
}
\description{
Create a table of the cytotoxicity data for the given cyto_type (AB or LDH) for all plates in a given input cytoFile
}
\details{
This function splits the cyto data for each assay and plate, then calls createCytoData() to read the data
}
