% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_conversion.R
\name{h5_conversion}
\alias{h5_conversion}
\title{Convert spike list files to h5 files}
\usage{
h5_conversion(
  project.output.dir,
  files.log.output.dir = project.output.dir,
  remake_all = FALSE,
  check_nwells_per_plate = 48,
  recording_duration_sec = 900
)
}
\arguments{
\item{project.output.dir}{folder in which the h5files should be created}

\item{files.log.output.dir}{folder in which the files_log is located
(i.e., the .txt file containing the list of spike list files and MaestroExperimentLog files to be read)}

\item{remake_all}{binary value indicating whether new h5files should be created
for every spike list file in the files_log (TRUE) or only for the spike list
files that have not already been created in the project.output.dir h5files folder (FALSE)}

\item{check_nwells_per_plate}{the number of wells that are expected to be
present in a plate (usually 48 wells for the NFA). If this many wells are not
found in the MaestroExperimentLog, h5_conversion will throw an error}

\item{recording_duration_sec}{expected duration of the spike list file
recording (in seconds). This should be 900 (i.e., 15 minutes) for Network
Formation Assay experiments done in the Shafer lab. Spikes recorded more than
recording_duration_sec beyond the time of the first spike will be ignored. If
a recording was more than 3 minutes less than the recording_duration_sec, the
function will throw an error.}
}
\description{
Convert spike list files to h5 files
}
\details{
convert_spkListFiles_to_h5_MetaData.R.
Created by Diana Hall. Adapted by Amy Carpenter
}
