% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_interpolate_DIV.R
\name{linear_interpolate_DIV}
\alias{linear_interpolate_DIV}
\title{Impute parameter values for standard DIV from non-standard DIV recordings using linear interpolation}
\usage{
linear_interpolate_DIV(dati, DIV.to.interpolate)
}
\arguments{
\item{dati}{data table in long format (i.e., 1 parameter per row)}

\item{DIV.to.interpolate}{desired DIV to interpolate from the closest DIV in dati}
}
\description{
Impute parameter values for standard DIV from non-standard DIV recordings using linear interpolation
}
\details{
This function finds the approximate value between 2 other values. NA values are set to 0 in order to find the slope
Last updated July 2023 by Amy Carpenter
}
