% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_values_to_AUC.R
\name{parameter_values_to_AUC}
\alias{parameter_values_to_AUC}
\title{Calculate Area-Under-the-Curve of parameter values vs DIV in each well}
\usage{
parameter_values_to_AUC(
  project.output.dir,
  project_name,
  div_data_file,
  id.columns = c("date", "Plate.SN", "DIV", "well", "trt", "dose", "units", "file.name",
    "wllq_by_well", "wllq_notes_by_well", "wllq_ref_by_well"),
  expected_DIVs = c(2, 5, 7, 9, 12),
  dose_column_name = NULL
)
}
\arguments{
\item{project.output.dir}{folder where the output folder should be saved}

\item{project_name}{name of the project (used to name the output file)}

\item{div_data_file}{full file path to single, wide .csv files containing the parameter data by DIV (created by prepare_parameter_values())}

\item{id.columns}{vector of column names in div_data_file that should be interpreted as ID columns (i.e., all columns other than the parameters)}

\item{expected_DIVs}{vector of the DIVs that should be present for all plates in div_data_file (defaults to 2, 5, 7, 9, 12). Used to check for missing DIV.}

\item{dose_column_name}{(optional) name of column in id.columns that contains the numeric dose. Will be rounded to 5 significant figures and converted to character to ensure consistency in representation of numeric value.}
}
\description{
Original from Chris Frank - February-March 2016
Adapted by Mahmoud 2019
Last edited by Amy Carpenter July 2023
Legacy script name: burst_parameter_to_AUC.R
}
