% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_parameter_values.R
\name{prepare_parameter_values}
\alias{prepare_parameter_values}
\title{Clean tables of parameter values by DIV and prepare for AUC calculation}
\usage{
prepare_parameter_values(
  project.output.dir,
  project_name,
  div_data_files,
  id.columns = c("date", "Plate.SN", "DIV", "well", "trt", "dose", "units", "file.name"),
  wllq.tb.by.well.file = NULL,
  num_rows_per_plate = 6,
  num_columns_per_plate = 8,
  expected_DIVs = c(5, 7, 9, 12),
  add_DIV2_values_of_0 = TRUE,
  interpolate_stnd_DIVs = TRUE
)
}
\arguments{
\item{project.output.dir}{folder where the output folder should be saved}

\item{project_name}{name of the project (used to name the output file)}

\item{div_data_files}{full file paths to wide .csv files containing the parameter data by DIV (with parameter values as column name(s))}

\item{id.columns}{vector of column names in div_data_files that should be interpreted as ID columns (i.e., all columns other than the parameters)}

\item{wllq.tb.by.well.file}{(optional) full path to table containing well quality definitions and/or notes for specific plates, wells, and DIVs. If NULL, wllq_by_well will be set to 1 for all data rows}

\item{num_rows_per_plate}{(optional) number of rows in an MEA plate. Used to expand data rows where well = 'all' in the wllq.tb.by.well. Defaults to 6 for a standard MEA 48-well plate.}

\item{num_columns_per_plate}{(optional) number of columns in an MEA plate. Used to expand data rows where well = 'all' in the wllq.tb.by.well. Defaults to 8 for a standard MEA 48-well plate.}

\item{expected_DIVs}{vector of the DIVs that should be present for all plates in div_data_files (defaults to 5, 7, 9, 12). Used to expands data rows where DIV = 'all' in wllq.tb.by.well and to check for non-standard DIV (see interpolate_stnd_DIVs)}

\item{add_DIV2_values_of_0}{whether to add dummy-data values of 0 at DIV 2 (standard practice for the Shafer Lab). If TRUE, the function will add 2 to expected_DIVs}

\item{interpolate_stnd_DIVs}{whether to interpolate standard DIVs (defined by expected_DIVs) from non-standard DIVs. Only applies if non-standard DIVs are present in the div_data_files}
}
\description{
This function reads the wide parameter data files in \code{div_data_files}, merges them, defines the wllq_by_well, and performs other optional checks and clean up.
The resulting single, wide data file containing all parameter values and DIV is saved as a csv in a folder named "output".
}
