% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcpl_MEA_dev_AUC.R
\name{tcpl_MEA_dev_AUC}
\alias{tcpl_MEA_dev_AUC}
\title{Read in wide-format AUC, DIV, and cytotoxicity data files and transform into 1 long-format data table}
\usage{
tcpl_MEA_dev_AUC(
  project.output.dir,
  project_name,
  assay_component_map_filename,
  auc_data_file = file.path(project.output.dir, "output", paste0(project_name,
    "_AUC.csv")),
  div_data_file = file.path(project.output.dir, "output", paste0(project_name,
    "_parameters_by_DIV.csv")),
  cytotox_data_file = file.path(project.output.dir, "output", paste0(project_name,
    "_cytotox.csv")),
  id.columns = c("date", "Plate.SN", "DIV", "well", "trt", "dose", "units", "file.name",
    "wllq_by_well", "wllq_notes_by_well", "wllq_ref_by_well")
)
}
\arguments{
\item{project.output.dir}{folder where the AUC, DIV, and cytotoxicity data files are located}

\item{project_name}{name of the project}

\item{assay_component_map_filename}{file path to the csv file containing a map of the assay component source names (acsn, aka "parameters") names as they appear in the AUC, DIV, and cytotoxicity data files to the corresponding assay component names (acnm) registered in invitroDB.
Note that this function will append 'DIV#" to the parameter values in the input DIV table in order to match with the names in the assay_component_map_filename.}

\item{auc_data_file}{(optional) file path to the csv file containing the AUC parameter values (created by prepare_parameter_values.R). Defaults to the AUC data file in the project.out.dir/output folder.}

\item{div_data_file}{(optional) file path to the csv file containing the parameter values by DIV (created by parameter_values_to_AUC.R). Defaults to the _parameters_by_DIV data file in the project.out.dir/output folder.}

\item{cytotox_data_file}{(optional) file path to the csv file containing the AUC parameter values (created by run_cytotox_functions.R). Defaults to the _cytotox data file in the project.out.dir/output folder.}

\item{id.columns}{vector of column names in auc_data_file, div_data_file, and cytotox_data_file that should be interpreted as ID columns (i.e., all columns other than the parameters).
This should include all columns that appear as an ID column in any of the 3 data files (any extra columns that do not apply to a particular data file will be ignored)}
}
\description{
Read in wide-format AUC, DIV, and cytotoxicity data files and transform into 1 long-format data table
}
